🔧 Replace `mypy` with `ty` (#691)
## Description

This PR replaces `mypy` with `ty` and fixes all found errors.

## Checklist

- [x] The pull request only contains commits that are focused and
relevant to this change.
- [x] ~~I have added appropriate tests that cover the new/changed
functionality.~~
- [x] ~~I have updated the documentation to reflect these changes.~~
- [x] The changes follow the project's style guidelines and introduce no
new warnings.
- [x] The changes are fully tested and pass the CI checks.
- [x] I have reviewed my own code changes.

+++ src/mqt/qecc/analog_information_decoding/simulators/simulation.py
@@ -475,7 +475,7 @@ class SingleShotSimulator:
         syndrome_w_err_binary += meta_decoded  # repair syndrome
         syndrome_w_err_binary %= 2
 
-        return syndrome_w_err_binary
+        return syndrome_w_err_binary.astype(np.int32)
 
     def _analog_tg_decoding(
         self,
@@ -709,7 +709,7 @@ class SingleShotSimulator:
         }
 
         output.update(self.input_values)
-        output["bias"] = replace_inf(output["bias"])  # type: ignore[assignment, arg-type]
+        output["bias"] = replace_inf(output["bias"])  # ty: ignore[invalid-assignment, invalid-argument-type]
 
         with Path(self.outfile).open(encoding=locale.getpreferredencoding(False)) as f:
             f.write(
